pyproject.toml
[tool.poetry]
name = "pymupdf_zeal_doc"
version = "0.1.1"
description = ""
authors = ["PowerSnail <hj@powersnail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
doc2dash = "^2.3.0"
Sphinx = "^4.1.2"
docopt-ng = "^0.7.2"
PyMuPDF = "^1.20.2"
docutils = "<0.16"
rst2pdf = "*"
sphinx-rtd-theme = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
